Move SharedMemData struct from AssembleElemSolverAlgorithm
to the SharedMemData header.
