+++ src/AssembleElemSolverAlgorithm.C
@@ -31,6 +31,7 @@
 
 #include <KokkosInterface.h>
 #include <ScratchViews.h>
+#include <SharedMemData.h>
 #include <SimdInterface.h>
 #include <CopyAndInterleave.h>
 
@@ -68,37 +69,6 @@ AssembleElemSolverAlgorithm::initialize_connectivity()
 {
   eqSystem_->linsys_->buildElemToNodeGraph(partVec_);
 }
-
-struct SharedMemData {
-    SharedMemData(const sierra::nalu::TeamHandleType& team,
-         const stk::mesh::BulkData& bulk,
-         const ElemDataRequests& dataNeededByKernels,
-         unsigned nodesPerEntity,
-         unsigned rhsSize)
-     : simdPrereqData(team, bulk, nodesPerEntity, dataNeededByKernels)
-    {
-        for(int simdIndex=0; simdIndex<simdLen; ++simdIndex) {
-          prereqData[simdIndex] = std::unique_ptr<ScratchViews<double> >(new ScratchViews<double>(team, bulk, nodesPerEntity, dataNeededByKernels));
-        }
-        simdrhs = get_shmem_view_1D<DoubleType>(team, rhsSize);
-        simdlhs = get_shmem_view_2D<DoubleType>(team, rhsSize, rhsSize);
-        rhs = get_shmem_view_1D<double>(team, rhsSize);
-        lhs = get_shmem_view_2D<double>(team, rhsSize, rhsSize);
-
-        scratchIds = get_int_shmem_view_1D(team, rhsSize);
-        sortPermutation = get_int_shmem_view_1D(team, rhsSize);
-    }
-
-    std::unique_ptr<ScratchViews<double>> prereqData[simdLen];
-    ScratchViews<DoubleType> simdPrereqData;
-    SharedMemView<DoubleType*> simdrhs;
-    SharedMemView<DoubleType**> simdlhs;
-    SharedMemView<double*> rhs;
-    SharedMemView<double**> lhs;
-
-    SharedMemView<int*> scratchIds;
-    SharedMemView<int*> sortPermutation;
-};
 
 //--------------------------------------------------------------------------
 //-------- execute ---------------------------------------------------------
